requirements.txt
paho-mqtt
apscheduler
requests
pymongo